pyproject.toml
[tool.poetry]
name = "zarrwlr"
version = "0.1.0"
description = "Python module/library for summarizing environmental sounds (or/and images) and metadata for scientific analysis by using Zarr databases."
authors = ["fherb2 <herbrand@gmx.de>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "zarrwlr", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.12.dev0"
mutagen = "^1.47.0"
rich = "^14.0.0"
zarr = "^3.0.7"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.12"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
pymdown-extensions = "^10.15"
mkdocs-section-index = "^0.3.10"
types-pyyaml = "^6.0.12.20250402"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.12"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
pymdown-extensions = "^10.15"
mkdocs-section-index = "^0.3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
